Use model object as central member in model ETL
The central member the _JDomModelETL_ must focus on is the _JDomModel_.
It now ensures that the model has been extracted before it is used,
transformed and/or loaded and that it's only loaded once (per ETL
instance). The model getter isn't deprecated anymore as we want to allow
it to be modified without requiring the use of the _transform_ method.

The _MavenProject_ member is not needed in the ETL anymore and thus has
been removed.