Improve insertion of new elements and whitespaces
New elements are not simply added to the end of the POM. They are now
inserted at a position that is defined using the field _elementOrder_
in the _JDomCfg_ class.

Furthermore, adding linebreaks and indentations has been unified. both
are now prepended to every new element that is inserted (using a JDOM
_Text_ object).