Improve tests that set JDom model beans
A number of integration tests created JDom model beans to verify that
not only the standard (Maven) model beans can be set (or added), but
also the JDom model beans that are introduced with this project. They
were actually meant to verify that POM elements read from another POM
can be set or added. The previous implementations of (some of) these
tests were quite cumbersome and hard to review.

The base class of the integration tests now allows to add a _source_
POM file where elements to be set or added to the subject POM can be
read from. With this, implementations of such tests are much easier to
implement and to understand now.

To allow easier identification of a plugin to be read from the _source_
POM, the implementation of the _JDomBuild_ class was changed to always
set the _plugins_ member of its parent class. This allows the lookup of
a requested plugin via _getPluginsAsMap_.