[#2] Set base model fields of JDomScm
The fields of the _JDomScm_s super class are now properly set when the
POM document is read or model members are changed. This provides easier
debugging and allows further (private) functionality of the super class
to be used.

Some new SCM integration tests were add to verify set and replace JDOM
SCM.