Fix addDependency method with add index
The previous implementation of the method
`JDomDependencies#addDependency(int index, Dependency dependency)`
only worked when the index was zero. When adding a _dependency_
at a desired position, the index of the JDOM element to be added
must be recalculated, respecting that between all elements there's
normally a whitespace text element containing line breaks and
indentations.

The integration test for that feature has been improved and moved to the
_DependenciesEtlIT_ class so that it now verifies the behaviour not only
for the _dependencyManagement_, but also for the plain _dependencies_.