Make compile with Java 11
This is questionable, though.  We override Properties#storeToXML by throwing an UnsupportedOperationException and ensure this by a test.  Why is this method considered so harmful that we definitely want to disable it?  With Java 11, you can easily bypass this limitation by using the Charset variant of storeToXML.  And we cannot prevent this, because as long as we are stuck to Java 8, we cannot override the Java 11 method.

CMS-18848